Moved data models to models package & other enh
